+++ ios/Classes/Bridge/StartStudioBridge.swift
@@ -27,7 +27,8 @@ struct StartStudioBridge: BaseBridge {
                     result(serialized)
 
                 case .error(let error):
-                    result(FlutterError(code: "error", message: error.localizedDescription, details: nil))
+                    let errorDetails = StartBridge.extractDetails(from: error)
+                    result(FlutterError(code: "error", message: error.localizedDescription, details: errorDetails))
 
                 case .cancel:
                     result(FlutterError(code: "exit", message: "User canceled the flow", details: nil))