+++ lib/src/model/model.dart
@@ -16,3 +16,4 @@ export 'package:onfido_sdk/src/model/cobranding_logo.dart';
 export 'package:onfido_sdk/src/model/ios_appearance.dart';
 export 'package:onfido_sdk/src/model/onfido_theme.dart';
 export 'package:onfido_sdk/src/model/background_color.dart';
+export 'package:onfido_sdk/src/model/nfc_options.dart';